FEAT: Created class for embedded figures
Created a class in plotting_utils for easily creating windows with embedded matplotlib figures that optionally have events. Created subclasses in peak_fitting for selecting peaks and selecting background points for doing peak fitting, and a subclass in peak_fitting_gui for a simple embedded figure in a window with no events.


Former-commit-id: d170bd4a5f8c2ba34f9d6236d7376547d4b0b64b