FIX: Fixed issue when using ConstantModel
Ensured that the background model for peak fitting is always an array with the same size as the data being fit so that it does not cause errors when plotting. Made the plotting convenience functions in peak_fitting take just the fit result, since the x and y values are within the result. Renamed the peak_fitting function to fit_peaks because it was causing a namespace issue.


Former-commit-id: 1be1e930eaa3ab490c8c850044e3854f9c0e9c3d