FEAT: File import options are more flexible
Made the file import options more flexible when used from the main gui by allowing user to select whether or not to use same import options for all files (option is disabled for Excel files). Also allow inputting previous inputs for the file import function, so that any changes to the defaults can be maintained, especially helpful when manually opening files for fitting or plotting.


Former-commit-id: b3688c8b7d9ef6b69fd7216de08fb161cb48b0fb