FEAT: Allow PreprocessingFunctions to delete columns
Renamed SeparationFunction to PreprocessingFunction to give it a more general meaning. Allow columns to be removed by PreprocessingFunctions, which update the unique variable references. If not processing, the column numbers for the deleted columns are removed from the DataSource before showing the import window. Made the string representation for DataSource and Function objects better and removed their repr. Added doc strings for everything in mcetl.functions.


Former-commit-id: 68e619ccd099dd4f5a8a9d4a7a588941067929dd