FIX: Check if xlrd is installed for reading .xls
Check whether xlrd is installed before adding .xls to the file search window. If xlrd is not installed, manually set the Excel reader engine to openpyxl so that no errors occur. Added xlrd to 'extras' for installation. Updated required pandas version to 0.24, which allows the engine keyword for read_excel, and lmfit to 1.0, since lower versions had issues.


Former-commit-id: 2533290be1de075099761bb14ccaa7fdbcdb3c22