FIX: Fixed matplotlib toolbar size and packing
In matplotlib version 3.3, the toolbar size was increased, so all toolbars in mcetl are likewise incrased in size. Likewise, a pack_toolbar keyword was added for NavigationToolbarTK, so try using that key when creating toolbars. Otherwise, use pack_forget to unpack the toolbar and then pack with the desired keys.