OTHER: Replace sympy with asteval
Use asteval instead of sympy to parse user expressions when creating secondary axes for the plotting gui. This requires the user to input forward and backward expressions, but otherwise requires no changes. Also, it now drops a requirement for mcetl, since asteval is already required for lmfit. Updated requirements in setup.py and requirements.txt.