MAINT: Bump required pandas version and drop optional requirements
Bump the required pandas version to 0.25 so that openpyxl can be used as an engine for pandas.read_excel. Removed the extras_require tag in setup.py, since users can just pip install other libraries. Made utils.check_requirements cache the results so that global constants do not need to be made to check for optional library availability.