OTHER: Simplified fit result processing to be name-independent
Simplified the processing of fit results in fitting_gui to be less coupled with the output of peak_fitting.fit_peaks, so that it will be easier to use with general fitting. peak_fitting.fit_peaks no longer does post-processing, so that users can use it rather than the fitting gui to get less processed results.