FIX: handle PermissionError for main gui
Safely handle PermissionErrors when deciding the folder to save previous_files json files, and when writing those files. This way, the program will not fail immediately upon importing mcetl if permission is denied in the desired folders so that user can change the folder if desired. Read access and checking if individual files in the folder still raise PermissionError so that user is notified that their folder is not usable and can change the folder.


Former-commit-id: b62f751bc351827739b4980794c915049d0a93da